pyproject.toml
[tool.poetry]
name = "users_and_posts"
version = "0.1.0"
description = "Test task for star-navi"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^4.7.1"
psycopg2 = "^2.9.1"
aiohttp = "^3.7.4"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
